+++ Protocol/Bingo/Aristocrat.Monaco.Bingo/CompositionRoot/InternalServiceBuilder.cs
@@ -53,10 +53,10 @@
             container.RegisterSingleton<IReportEventQueueService, ReportEventHandler>();
             container.RegisterSingleton<IGameHistoryReportHandler, GameHistoryReportHandler>();
             container.RegisterSingleton<ISharedConsumer, SharedConsumerContext>();
-            container.RegisterSingleton<IAcknowledgedQueue<ReportTransactionMessage, int>, AcknowledgedQueue<ReportTransactionMessage, int>>();
-            container.RegisterSingleton<IAcknowledgedQueueHelper<ReportTransactionMessage, int>, TransactionAcknowledgedQueueHelper>();
-            container.RegisterSingleton<IAcknowledgedQueue<ReportEventMessage, int>, AcknowledgedQueue<ReportEventMessage, int>>();
-            container.RegisterSingleton<IAcknowledgedQueueHelper<ReportEventMessage, int>, EventAcknowledgedQueueHelper>();
+            container.RegisterSingleton<IAcknowledgedQueue<ReportTransactionMessage, long>, AcknowledgedQueue<ReportTransactionMessage, long>>();
+            container.RegisterSingleton<IAcknowledgedQueueHelper<ReportTransactionMessage, long>, TransactionAcknowledgedQueueHelper>();
+            container.RegisterSingleton<IAcknowledgedQueue<ReportEventMessage, long>, AcknowledgedQueue<ReportEventMessage, long>>();
+            container.RegisterSingleton<IAcknowledgedQueueHelper<ReportEventMessage, long>, EventAcknowledgedQueueHelper>();
             container.RegisterSingleton<IAcknowledgedQueue<ReportGameOutcomeMessage, long>, AcknowledgedQueue<ReportGameOutcomeMessage, long>>();
             container.RegisterSingleton<IAcknowledgedQueueHelper<ReportGameOutcomeMessage, long>, GameHistoryReportAcknowledgeQueueHelper>();
             container.RegisterSingleton<IBingoGameProvider, BingoGameProvider>();
